REFACTOR samples.py to samples directory & moved output files
