--- samples/demo_encrypt.py
@@ -1,26 +1,0 @@
-# Test encrypt module reliability
-from pdfconduit import Encrypt, Info
-from tests import pdf_path
-
-
-def main():
-    print('Testing Encrypt reliability')
-    owner_pw = 'foo'
-    user_pw = 'baz'
-
-    p = Encrypt(pdf_path, user_pw, owner_pw)
-    print(p.output)
-
-    security = Info(p.output, user_pw).security
-
-    try:
-        assert Info(p.output, user_pw).encrypted is True
-        assert security['/Length'] == 128
-        assert security['/P'] == -1852
-        print('Success!')
-    except AssertionError:
-        print('Failed!')
-
-
-if __name__ == '__main__':
-    main()
