--- samples/demo_watermark_gui.py
@@ -1,9 +1,0 @@
-from pdfconduit import GUI
-
-
-def main():
-    GUI.watermark()
-
-
-if __name__ == '__main__':
-    main()
